discourse/pyproject.toml
[tool.poetry]
name = "discourse"
version = "0.1.0"
description = ""
authors = ["tomfutago <tomfutago@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
requests = "^2.28.1"
python-dotenv = "^0.20.0"
pandas = "^1.4.3"
SQLAlchemy = "^1.4.40"
psycopg2-binary = "^2.9.3"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
